TERR-53 - added reporters to karma config - refactored santize functions into utils module - narrowed scope of coverage testing to non-d3, non-leaflet, non-pubsub code - corrected imports of santize functions - refactored render function to use resolvePath - removed resolveObjPath fn from templates module - removed topologyTools module - added some jsdoc inlinedocs for utils module - refactored util fn name from getUrlSearchParams to getUrlQueryParams
