+++ index.html
@@ -13,7 +13,7 @@
     </script>
 
     <script type="module">
-        import { sanitizeTopology } from "./src/components/lib/topologyTools.js"
+        import { sanitizeTopology } from "./src/components/lib/utils.js"
         var parsedData = {};
         var lavender = "rgb(202, 149, 229)"
         var topology = [
